Problem: CI for gozyre fails
Solution: move set_beacon_port to `options_draft.go` hidden by build
tag, change CI config to install git-draft dependencies and test draft
build too
