+++ .circleci/config.yml
@@ -16,7 +16,7 @@ jobs: # basic units of work in a run
       - checkout # check out source code to working directory
       - run: 
         #TODO: create own Docker container under hub.docker.com/zeromq/
-          name: Add C dependencies
+          name: Add stable dependencies
           command: |
             sudo sed -e '$s,$,\ndeb http://download.opensuse.org/repositories/network:/messaging:/zeromq:/release-stable/Debian_10/ ./,' -i /etc/apt/sources.list
             sudo curl 'http://download.opensuse.org/repositories/network:/messaging:/zeromq:/release-stable/Debian_10/Release.key' > Release.key
@@ -37,12 +37,28 @@ jobs: # basic units of work in a run
       - run: go get github.com/jstemmer/go-junit-report
 
       - run:
-          name: Run unit tests
+          name: Run unit tests for stable version
           command: |
             trap "go-junit-report <${TEST_RESULTS}/go-test.out > ${TEST_RESULTS}/go-test-report.xml" EXIT
             go test | tee ${TEST_RESULTS}/go-test.out
 
-      - run: go build # pull and build dependencies for the project
+      - run: go build # pull and build dependencies for the projec
+
+      - run: 
+          name: Add draft dependencies
+          command: |
+            sudo apt-get remove libzyre-dev libzmq3-dev libczmq-dev libzyre2 libczmq4 libzmq5
+            sudo sed -e 's/release-stable/git-draft/' -i /etc/apt/sources.list
+            sudo curl 'http://download.opensuse.org/repositories/network:/messaging:/zeromq:/git-draft/Debian_10/Release.key' > Release.key
+            sudo apt-key add - < Release.key
+            sudo apt-get update
+            sudo apt-get install libzyre-dev libzmq3-dev libczmq-dev libzyre2 libczmq4 libzmq5
+
+      - run:
+          name: Run unit tests for draft version
+          command: |
+            trap "go-junit-report <${TEST_RESULTS}/go-test.out > ${TEST_RESULTS}/go-test-report.xml" EXIT
+            go test -tags=draft | tee ${TEST_RESULTS}/go-test.out
 
       - save_cache: # Store cache in the /go/pkg directory
           key: v1-pkg-cache
